+++ CHANGES.md
@@ -34,6 +34,7 @@ Changes in 13.0
 - Use RosarioSIS\Functions\Email class, program wide
 - Remove deprecated `_rosarioLoginURL()` function in SendNotification.fnc.php
 - Add `&modfunc=save` to Fields form in Fields.fnc.php, SchoolFields.php, AddressFields.php, PeopleFields.php, StudentFields.php & UserFields.php
+- Remove Debug.fnc.php file include: Kint() & d() functions in Warehouse.php
 
 Changes in 12.9.2
 -----------------